Pre-fill the output dict with the defaults
The dictionary output by the parse method is now filled with the
default values before it is filled with the user-provided ones. This
makes all the defined options available in the parser output, even if
they were not explicitly specified by the user.

In the case of "multi" options, en empty list is set as default if no
default is given (i.e default is None).