Fix the help generation
The display of the current values in the help string was relying on the
Option object being mutated at parsing time. With Options not being
updated anymore, the default values where displayed.

This commit introduces a help method that takes the command line
arguments as argument. This method parses the cli args to fill the
values in the help.

Note that print(OPTION) will display the default values if the Options
instance is not initialized with args. use
print(OPTION.help(sys.argv[1:])) instead.